Remove "Your account" section
This removes a section from the app that currently allows users to
change the details of their account, but users will be signing in via
CIS2 in the future so we don't need this functionality.
