+++ app/views/dashboard/index.html.erb
@@ -37,9 +37,3 @@
     <% end %>
   </li>
 </ul>
-
-<h2 class="nhsuk-heading-s nhsuk-u-margin-bottom-0">
-  <%= link_to "Your account", users_account_path(current_user),
-              class: "nhsuk-link--no-visited-state" %>
-</h2>
-<p>Change your email address or name.</p>
